Fix integer type to float

+++ game/base_strategy.py
@@ -16,7 +16,7 @@ class BaseStrategy:
         '''
         raise NotImplementedError("Must implement select_planes method!")
 
-    def steer_input(self, planes: dict[str, Plane]) -> dict[str, int]:
+    def steer_input(self, planes: dict[str, Plane]) -> dict[str, float]:
         '''
         Return a dictionary mapping each plane id to the amount they will steer [-1, 1], where positive is clockwise
         '''
